refactor(discord): extract sweeper abandon guard
Keep the giant placeholder sweeper below its frozen baseline while retaining
fail-closed abandoned-turn cleanup and its focused policy tests.
