+++ src/services/discord/placeholder_sweeper.rs
@@ -30,13 +30,17 @@ use super::formatting::{
 };
 use super::gateway::edit_outbound_message;
 use super::inflight::{
-    DEAD_WATCHER_PROVEN_DEAD_SECS, InflightTurnState, delete_inflight_state_file,
-    emit_reap_abandoned_rebind_origin, load_inflight_states_for_sweep, parse_started_at_unix,
-    reap_abandoned_rebind_origin_locked, should_reap_abandoned_rebind_origin,
-    sweep_reap_dead_watcher_rebind_origin,
+    InflightTurnState, delete_inflight_state_file, emit_reap_abandoned_rebind_origin,
+    load_inflight_states_for_sweep, parse_started_at_unix, reap_abandoned_rebind_origin_locked,
+    should_reap_abandoned_rebind_origin, sweep_reap_dead_watcher_rebind_origin,
 };
-use crate::services::platform::tmux::PaneLiveness;
 use crate::services::provider::ProviderKind;
+
+mod abandon_guard;
+use abandon_guard::{
+    AbandonedTmuxCleanupDecision, abandoned_tmux_cleanup_decision_for,
+    finalize_abandoned_mailbox_if_proven_dead,
+};
 
 /// Age (seconds since `updated_at`) at which a placeholder is treated as
 /// stalled. Below this threshold the sweeper does nothing.
@@ -101,71 +105,6 @@ fn classify_age(age_secs: u64) -> SweepDecision {
     } else {
         SweepDecision::Active
     }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum RuntimeActivityEvidence {
-    Recent,
-    Inactive,
-    Unknown,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum AbandonedTmuxCleanupDecision {
-    Kill,
-    NoKillMarkerOnly,
-}
-
-fn abandoned_tmux_cleanup_decision(
-    has_usable_session_name: bool,
-    pane: PaneLiveness,
-    activity: RuntimeActivityEvidence,
-) -> AbandonedTmuxCleanupDecision {
-    if has_usable_session_name
-        && pane == PaneLiveness::DeadOrAbsent
-        && activity == RuntimeActivityEvidence::Inactive
-    {
-        AbandonedTmuxCleanupDecision::Kill
-    } else {
-        AbandonedTmuxCleanupDecision::NoKillMarkerOnly
-    }
-}
-
-fn runtime_activity_evidence(session_name: &str) -> RuntimeActivityEvidence {
-    let latest_nanos = crate::services::dispatched_sessions::latest_runtime_activity_unix_nanos(
-        session_name,
-    );
-    if latest_nanos <= 0 {
-        return RuntimeActivityEvidence::Unknown;
-    }
-    let latest_secs = latest_nanos / 1_000_000_000;
-    let now_secs = chrono::Utc::now().timestamp();
-    let age_secs = now_secs.saturating_sub(latest_secs).max(0) as u64;
-    if age_secs < DEAD_WATCHER_PROVEN_DEAD_SECS {
-        RuntimeActivityEvidence::Recent
-    } else {
-        RuntimeActivityEvidence::Inactive
-    }
-}
-
-fn abandoned_tmux_cleanup_decision_for(
-    state: &InflightTurnState,
-) -> AbandonedTmuxCleanupDecision {
-    if state.user_msg_id == 0 {
-        return AbandonedTmuxCleanupDecision::NoKillMarkerOnly;
-    }
-    let Some(session_name) = state.tmux_session_name.as_deref() else {
-        return AbandonedTmuxCleanupDecision::NoKillMarkerOnly;
-    };
-    let session_name = session_name.trim();
-    if session_name.is_empty() {
-        return AbandonedTmuxCleanupDecision::NoKillMarkerOnly;
-    }
-    abandoned_tmux_cleanup_decision(
-        true,
-        crate::services::tmux_diagnostics::tmux_session_pane_liveness(session_name),
-        runtime_activity_evidence(session_name),
-    )
 }
 
 fn build_stalled_placeholder(state: &InflightTurnState) -> String {
@@ -770,49 +709,6 @@ impl StalledEditTracker {
     }
 }
 
-/// Finalize and cancel an abandoned turn only after independent tmux and
-/// runtime-activity evidence proves its owner is gone. Every other outcome is
-/// marker-only so a live or indeterminate session remains untouched.
-async fn finalize_abandoned_mailbox_if_proven_dead(
-    shared: &Arc<SharedData>,
-    provider: &ProviderKind,
-    state: &InflightTurnState,
-) -> bool {
-    if abandoned_tmux_cleanup_decision_for(state) != AbandonedTmuxCleanupDecision::Kill {
-        return false;
-    }
-
-    let channel = serenity::ChannelId::new(state.channel_id);
-    let finish = super::mailbox_finish_turn_if_matches(
-        shared,
-        provider,
-        channel,
-        serenity::MessageId::new(state.user_msg_id),
-    )
-    .await;
-    if let Some(removed_token) = finish.removed_token {
-        super::turn_bridge::cancel_active_token(
-            &removed_token,
-            super::TmuxCleanupPolicy::CleanupSession {
-                termination_reason_code: Some("placeholder_sweeper_abandon"),
-            },
-            "placeholder_sweeper abandoned",
-        );
-        super::saturating_decrement_global_active(shared);
-        if finish.has_pending {
-            super::schedule_deferred_idle_queue_kickoff(
-                shared.clone(),
-                provider.clone(),
-                channel,
-                "placeholder_sweeper_abandon",
-            );
-        }
-        true
-    } else {
-        false
-    }
-}
-
 /// True when the inflight state on disk for `state.channel_id` still names
 /// the same turn (matching `user_msg_id` and `current_msg_id`) AND the file
 /// mtime is not significantly fresher than our snapshot. Returns `false`
@@ -1326,11 +1222,9 @@ mod safety_net_threshold_tests {
     //! explicit-signal wire (D / A / B / C) must have time to fire
     //! before the sweeper does anything destructive.
     use super::{
-        ABANDON_THRESHOLD_SECS, AbandonedTmuxCleanupDecision, INITIAL_DELAY_SECS,
-        RuntimeActivityEvidence, STALL_THRESHOLD_SECS, SWEEP_INTERVAL_SECS,
-        abandoned_tmux_cleanup_decision, panel_reclaim_target,
+        ABANDON_THRESHOLD_SECS, INITIAL_DELAY_SECS, STALL_THRESHOLD_SECS, SWEEP_INTERVAL_SECS,
+        panel_reclaim_target,
     };
-    use crate::services::platform::tmux::PaneLiveness;
     use crate::services::provider::ProviderKind;
 
     fn sweep_state_with_panel(status_message_id: Option<u64>) -> super::InflightTurnState {
@@ -1407,78 +1301,5 @@ mod safety_net_threshold_tests {
         let target = panel_reclaim_target(&state, ABANDON_THRESHOLD_SECS);
 
         assert_eq!(target, None);
-    }
-
-    #[test]
-    fn no_death_evidence_keeps_the_tmux_session() {
-        assert_eq!(
-            abandoned_tmux_cleanup_decision(
-                true,
-                PaneLiveness::DeadOrAbsent,
-                RuntimeActivityEvidence::Unknown,
-            ),
-            AbandonedTmuxCleanupDecision::NoKillMarkerOnly,
-        );
-    }
-
-    #[test]
-    fn missing_or_blank_tmux_name_is_marker_only_without_a_probe() {
-        assert_eq!(
-            abandoned_tmux_cleanup_decision(
-                false,
-                PaneLiveness::DeadOrAbsent,
-                RuntimeActivityEvidence::Inactive,
-            ),
-            AbandonedTmuxCleanupDecision::NoKillMarkerOnly,
-        );
-    }
-
-    #[test]
-    fn panel_only_state_is_marker_only() {
-        let mut state = sweep_state_with_panel(Some(5001));
-        state.user_msg_id = 0;
-
-        assert_eq!(
-            super::abandoned_tmux_cleanup_decision_for(&state),
-            AbandonedTmuxCleanupDecision::NoKillMarkerOnly,
-        );
-    }
-
-    #[test]
-    fn confirmed_dead_pane_with_confirmed_inactivity_allows_cleanup() {
-        assert_eq!(
-            abandoned_tmux_cleanup_decision(
-                true,
-                PaneLiveness::DeadOrAbsent,
-                RuntimeActivityEvidence::Inactive,
-            ),
-            AbandonedTmuxCleanupDecision::Kill,
-        );
-    }
-
-    #[test]
-    fn live_pane_keeps_the_tmux_session_even_when_activity_is_not_recent() {
-        assert_eq!(
-            abandoned_tmux_cleanup_decision(
-                true,
-                PaneLiveness::Live,
-                RuntimeActivityEvidence::Inactive,
-            ),
-            AbandonedTmuxCleanupDecision::NoKillMarkerOnly,
-        );
-    }
-
-    #[test]
-    fn ambiguous_pane_and_activity_evidence_is_marker_only() {
-        for (pane, activity) in [
-            (PaneLiveness::ProbeError, RuntimeActivityEvidence::Inactive),
-            (PaneLiveness::DeadOrAbsent, RuntimeActivityEvidence::Recent),
-            (PaneLiveness::Live, RuntimeActivityEvidence::Unknown),
-        ] {
-            assert_eq!(
-                abandoned_tmux_cleanup_decision(true, pane, activity),
-                AbandonedTmuxCleanupDecision::NoKillMarkerOnly,
-            );
-        }
     }
 }
